Reorganise webpack configuration + fix demo
