+++ src/demo/index.html
@@ -5,6 +5,6 @@
     </head>
     <body>
         <div id='root'></div>
-        <script src="/dash_loading_spinners/output.js"></script>
+        <script src="/demo-lib/dash_loading_spinners.js"></script>
     </body>
 </html>
