+++ webpack/moduleDefinition.js
@@ -1,0 +1,25 @@
+module.exports = {
+  rules: [
+    {
+      test: /\.jsx?$/,
+      exclude: /node_modules/,
+      use: {
+        loader: 'babel-loader',
+      },
+    },
+    {
+      test: /\.css$/,
+      use: [
+        {
+          loader: 'style-loader',
+          options: {
+            insertAt: 'top',
+          },
+        },
+        {
+          loader: 'css-loader',
+        },
+      ],
+    },
+  ],
+};
